requirements.txt
Pillow==10.4.0
torch==2.3.1
transformers==4.43.3
arabert==1.0.1
boto3
awscli